Fix issues with eye position and view frustrum.
Having the eye position directly on top of the "body" made it possible
for the eye to be directly touching the bottom of a cube.  This made a
reasonable znear/zfar impossible.  Introduced a forehead and chose values
for znear/zfar that avoid seeing through nearby objects without making
the Z buffer resolution too low.